switched to using isFolder, and moved it from cms-cli to cms-lib, updating refs
