+++ packages/cms-cli/lib/filesystem.js
@@ -1,6 +1,5 @@
 const path = require('path');
 const { getCwd } = require('@hubspot/cms-lib/path');
-const { FOLDER_DOT_EXTENSIONS } = require('@hubspot/cms-lib/lib/constants');
 
 function resolveLocalPath(filepath) {
   return filepath && typeof filepath === 'string'
@@ -9,22 +8,6 @@ function resolveLocalPath(filepath) {
       getCwd();
 }
 
-function isPathFolder(path) {
-  const splitPath = path.split('/');
-  const fileOrFolderName = splitPath[splitPath.length - 1];
-  const splitName = fileOrFolderName.split('.');
-
-  if (
-    splitName.length > 1 &&
-    FOLDER_DOT_EXTENSIONS.indexOf(splitName[1]) === -1
-  ) {
-    return false;
-  }
-
-  return true;
-}
-
 module.exports = {
   resolveLocalPath,
-  isPathFolder,
 };
